added draw

+++ Classes.py
@@ -2,7 +2,7 @@ from math import *
 G = 6.67430e-11
 
 class Astre:
-    def __init__(self, rayon: float, masse: float, x: int, y: int) -> None:
+    def __init__(self, rayon: int, masse: float, x: int, y: int) -> None:
         self.r = rayon
         self.m = masse
         self.x = x
